requirements.txt
matplotlib==3.3.2
numpy==1.19.1
cvxpy==1.1.5
torch==0.4.1.post2
PyYAML==5.3.1
tensorboardX==2.1
